Cache log formatter for LoggerImplementation
Fixes: https://sentry.io/swiftypecom/work/issues/622966103/

The Rails logger in website uses `ActiveSupport::TaggedLogging` which
extends both the logger and the formatter. `Loops::Logger#logfile=` is
called multiple times during the startup process, once before Rails
configuration and once after. The second call creates another instance
of `LoggerImplementation` with a new formatter that isn't extended with
the tagged logging methods.

This caches the formatter on the parent logger instance and passes it to
the implementation so that the object will only be created once.